pyproject.toml
[tool.poetry]
name = "paddleocr-cli"
version = "0.1.0"
description = "A command-line tool for OCR using PaddleOCR AI Studio API"
authors = ["AI Interview Assistant"]
license = "MIT"
readme = "README.md"
packages = [{include = "paddleocr_cli"}]
keywords = ["ocr", "paddleocr", "cli", "document", "pdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
    "Topic :: Text Processing",
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = {version = "^6.0", optional = true}

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.0"
ruff = "^0.4"
mypy = "^1.10"

[tool.poetry.scripts]
paddleocr_cli = "paddleocr_cli.cli:main"
paddleocr-cli = "paddleocr_cli.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=paddleocr_cli"
